src/main/java/com/mjkam/openseaapiserver/util/BlockchainUtils.java
package com.mjkam.openseaapiserver.util;

import org.web3j.crypto.ECDSASignature;
import org.web3j.crypto.Hash;
import org.web3j.crypto.Keys;
import org.web3j.crypto.Sign;
import org.web3j.utils.Numeric;

import java.math.BigInteger;
import java.nio.charset.StandardCharsets;
import java.util.Arrays;

public class BlockchainUtils {
    private static final String MESSAGE_PREFIX = "\u0019Ethereum Signed Message:\n";

    public static boolean isRightSigner(String signer, String message, String signature) {
        try {
            String recoveredSigner = recoverAddress(BlockchainUtils.hashMessage(message), signature);
            if(recoveredSigner == null) {
                throw new VerifySignatureException("recoveredSigner is null");
            }
            return recoveredSigner.equals(signer);
        } catch (Exception e) {
            throw new VerifySignatureException("verify failed", e);
        }
    }

    private static String hashMessage(String message) {
        return Hash.sha3(
                Numeric.toHexStringNoPrefix(
                        (BlockchainUtils.MESSAGE_PREFIX + message.length() + message).getBytes(StandardCharsets.UTF_8)));
    }

    private static String recoverAddress(String digest, String signature) {
        Sign.SignatureData signatureData = BlockchainUtils.getSignatureData(signature);
        int header = 0;
        for (byte b : signatureData.getV()) {
            header = (header << 8) + (b & 0xFF);
        }
        if (header < 27 || header > 34) {
            return null;
        }
        int recId = header - 27;

        BigInteger key = Sign.recoverFromSignature(
                recId,
                new ECDSASignature(new BigInteger(1, signatureData.getR()), new BigInteger(1, signatureData.getS())),
                Numeric.hexStringToByteArray(digest));
        if (key == null) {
            return null;
        }
        return ("0x" + Keys.getAddress(key)).trim();
    }

    private static Sign.SignatureData getSignatureData(String signature) {
        byte[] signatureBytes = Numeric.hexStringToByteArray(signature);
        byte v = signatureBytes[64];
        if (v < 27) {
            v += 27;
        }
        byte[] r = Arrays.copyOfRange(signatureBytes, 0, 32);
        byte[] s = Arrays.copyOfRange(signatureBytes, 32, 64);
        return new Sign.SignatureData(v, r, s);
    }
}